requirements.txt
Flask==3.0.3
Flask-Cors==6.0.1

langchain==0.2.11
langchain-core==0.2.36
langchain-google-genai==1.0.10

python-dotenv==1.0.1

gunicorn==20.1.0
langchain-community

pymongo==4.8.0
pydantic==2.9.2
typing-extensions==4.12.2

pytz
